[AOTI] Extend torchgen to generate C shim with version number
Summary: While it is ok to add a new arg with defaul value to a fallback op in Python, it will be BC-breaking for the C shim. This PR adds an automatic approach to update C shim files when specifying a version number with a list of new args for the modified op. TO-BE-FILLED: there will be an example PR linked here later.

ghstack-source-id: 25ad2c8543b8a97d8f680a3efd5903e31bbf3628
Pull Request resolved: https://github.com/pytorch/pytorch/pull/147745